+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -621,6 +621,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Cloned volume profile from read only assets are created in the root of the project. (case 1154961)
 - Fixed Wizard check on default volume profile to also check it is not the default one in package.
 - Fix erroneous central depth sampling in TAA.
+- Fixed light layers not correctly disabled when the lightlayers is set to Nothing and Lightlayers isn't enabled in HDRP Asset
 
 ### Changed
 - Improve MIP selection for decals on Transparents